requirements.txt
astroid==2.2.5
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
cloudinary==1.16.0
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2.1
django-nose==1.4.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mock==3.0.5
nose==1.3.7
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pytest==4.5.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
wrapt==1.11.1
